refactor: changed test setup and internal distribution
- Removed the embedded_protocol.c and embedded_protocol.h files as they are no longer needed, replaced with main.c
- Introduced connection_manager.c and connection_manager.h to handle server and client connections.
- Created msg_handler.c and msg_handler.h for message handling logic.
- Implemented a ThreadSafeQueue for managing task and message queues.
- Refactored task_manager.c to utilize the new queue system for task management.
- Updated sender and receiver logic to work with the new queue structure.
- Updated tests to reflect changes in the message queuing and handling mechanisms.
