+++ src/elasticai/experiment_framework/remote_control/embedded_side/connection_manager.h
@@ -1,4 +1,8 @@
-#include "frame.h"
+#ifndef CONNECTION_MANAGER_H
+#define CONNECTION_MANAGER_H
+
+#include <unistd.h>
+#include <stdint.h>
 
 typedef struct
 {
@@ -16,12 +20,6 @@ Client start_client(int port);
 
 void close_connection(int fd);
 
-void enqueue_message(SendOrder *order);
-SendOrder dequeue_message();
-
 void print_payload(uint8_t *payload, uint8_t payload_len);
 
-int send_frame(int socket, Frame *frame);
-Frame read_frame(int socket);
-
-void *sending_thread(void *arg);
+#endif
