+++ src/elasticai/experiment_framework/remote_control/embedded_side/frame.h
@@ -21,10 +21,4 @@ typedef struct
     uint8_t *payload;
 } Frame;
 
-typedef struct
-{
-    uint8_t fd;
-    Frame frame;
-} SendOrder;
-
 #endif
